Add visitor pattern to the triangle tree (#63333)
In order to add more functionality to this data structure, this commit adds a visitor pattern to the tree.
