+++ x-pack/plugin/spatial/src/test/java/org/elasticsearch/xpack/spatial/index/fielddata/GeoShapeCoordinateEncoderTests.java
@@ -8,9 +8,7 @@ package org.elasticsearch.xpack.spatial.index.fielddata;
 
 import org.apache.lucene.geo.GeoEncodingUtils;
 import org.elasticsearch.test.ESTestCase;
-import org.elasticsearch.xpack.spatial.index.fielddata.GeoShapeCoordinateEncoder;
 
-import static org.elasticsearch.xpack.spatial.index.fielddata.GeoShapeCoordinateEncoder.INSTANCE;
 import static org.hamcrest.Matchers.closeTo;
 import static org.hamcrest.Matchers.endsWith;
 import static org.hamcrest.Matchers.equalTo;
@@ -22,16 +20,16 @@ public class GeoShapeCoordinateEncoderTests extends ESTestCase {
         double randomInvalidLon = randomFrom(randomDoubleBetween(-1000, -180.01, true),
             randomDoubleBetween(180.01, 1000, true));
 
-        assertThat(INSTANCE.encodeX(Double.POSITIVE_INFINITY), equalTo(Integer.MAX_VALUE));
-        assertThat(INSTANCE.encodeX(Double.NEGATIVE_INFINITY), equalTo(Integer.MIN_VALUE));
-        int encodedLon = INSTANCE.encodeX(randomLon);
+        assertThat(CoordinateEncoder.GEO.encodeX(Double.POSITIVE_INFINITY), equalTo(Integer.MAX_VALUE));
+        assertThat(CoordinateEncoder.GEO.encodeX(Double.NEGATIVE_INFINITY), equalTo(Integer.MIN_VALUE));
+        int encodedLon = CoordinateEncoder.GEO.encodeX(randomLon);
         assertThat(encodedLon, equalTo(GeoEncodingUtils.encodeLongitude(randomLon)));
-        Exception e = expectThrows(IllegalArgumentException.class, () -> GeoShapeCoordinateEncoder.INSTANCE.encodeX(randomInvalidLon));
+        Exception e = expectThrows(IllegalArgumentException.class, () -> CoordinateEncoder.GEO.encodeX(randomInvalidLon));
         assertThat(e.getMessage(), endsWith("must be between -180.0 and 180.0"));
 
-        assertThat(INSTANCE.decodeX(encodedLon), closeTo(randomLon, 0.0001));
-        assertThat(INSTANCE.decodeX(Integer.MAX_VALUE), closeTo(180, 0.00001));
-        assertThat(INSTANCE.decodeX(Integer.MIN_VALUE), closeTo(-180, 0.00001));
+        assertThat(CoordinateEncoder.GEO.decodeX(encodedLon), closeTo(randomLon, 0.0001));
+        assertThat(CoordinateEncoder.GEO.decodeX(Integer.MAX_VALUE), closeTo(180, 0.00001));
+        assertThat(CoordinateEncoder.GEO.decodeX(Integer.MIN_VALUE), closeTo(-180, 0.00001));
     }
 
     public void testLatitude() {
@@ -39,15 +37,15 @@ public class GeoShapeCoordinateEncoderTests extends ESTestCase {
         double randomInvalidLat = randomFrom(randomDoubleBetween(-1000, -90.01, true),
             randomDoubleBetween(90.01, 1000, true));
 
-        assertThat(INSTANCE.encodeY(Double.POSITIVE_INFINITY), equalTo(Integer.MAX_VALUE));
-        assertThat(INSTANCE.encodeY(Double.NEGATIVE_INFINITY), equalTo(Integer.MIN_VALUE));
-        int encodedLat = INSTANCE.encodeY(randomLat);
+        assertThat(CoordinateEncoder.GEO.encodeY(Double.POSITIVE_INFINITY), equalTo(Integer.MAX_VALUE));
+        assertThat(CoordinateEncoder.GEO.encodeY(Double.NEGATIVE_INFINITY), equalTo(Integer.MIN_VALUE));
+        int encodedLat = CoordinateEncoder.GEO.encodeY(randomLat);
         assertThat(encodedLat, equalTo(GeoEncodingUtils.encodeLatitude(randomLat)));
-        Exception e = expectThrows(IllegalArgumentException.class, () -> GeoShapeCoordinateEncoder.INSTANCE.encodeY(randomInvalidLat));
+        Exception e = expectThrows(IllegalArgumentException.class, () -> CoordinateEncoder.GEO.encodeY(randomInvalidLat));
         assertThat(e.getMessage(), endsWith("must be between -90.0 and 90.0"));
 
-        assertThat(INSTANCE.decodeY(encodedLat), closeTo(randomLat, 0.0001));
-        assertThat(INSTANCE.decodeY(Integer.MAX_VALUE), closeTo(90, 0.00001));
-        assertThat(INSTANCE.decodeY(Integer.MIN_VALUE), closeTo(-90, 0.00001));
+        assertThat(CoordinateEncoder.GEO.decodeY(encodedLat), closeTo(randomLat, 0.0001));
+        assertThat(CoordinateEncoder.GEO.decodeY(Integer.MAX_VALUE), closeTo(90, 0.00001));
+        assertThat(CoordinateEncoder.GEO.decodeY(Integer.MIN_VALUE), closeTo(-90, 0.00001));
     }
 }
